Resolve ambiguous overload for ‘operator/’ in ‘data / m’

+++ src/io.cc
@@ -27,7 +27,7 @@ void dump(Data &data)
   static size_t frame = 0;
 
   const size_t m = 180;
-  const size_t n = data / m;
+  const size_t n = (size_t)data / m;
   const void  *h = data.host();
   const State *s = (State *)h;
